fix 1

+++ backend/controllers/ScheduleController/index.js
@@ -6,6 +6,7 @@ const transporter = require("../../helpers/mailer-config");
 const moment = require("moment");
 const Patient = require("../../models/Patient");
 const User = require("../../models/User");
+const Notification = require("../../models/Notification");
 require("moment/locale/vi");
 
 const createSchedule = async (req, res) => {
@@ -81,13 +82,33 @@ const deleteSchedule = async (req, res) => {
         .json({ success: false, message: "Schedule not found" });
     }
 
+    const doctor = await Doctor.findById(schedule.doctor_id);
+    if (!doctor) {
+      return res
+        .status(400)
+        .json({ success: false, message: "Doctor not found" });
+    }
+
+    const appointments = await Appointment.find({
+      doctor_id: doctor._id,
+      work_date: schedule.work_date,
+      work_shift: schedule.work_shift, 
+      status: "confirmed",
+    });
+    if (appointments.length > 0) {
+      return res
+        .status(400)
+        .json({ success: false, message: "Không thể xoá lịch làm việc khi có lịch hẹn!" });
+    }
+
     const now = new Date();
     const workDate = new Date(schedule.work_date);
     // Tính toán thời gian còn lại đến lịch làm việc
     const timeDifference = workDate - now;
 
-    // Kiểm tra nếu thời gian còn lại lớn hơn 24 giờ (24 * 60 * 60 * 1000 milliseconds)
+    // Kiểm tra nếu thời gian còn lại lớn hơn 24 giờ
     if (timeDifference > 24 * 60 * 60 * 1000) {
+      // Xóa lịch làm việc
       await Schedule.findByIdAndDelete(id);
       return res
         .status(200)
@@ -97,7 +118,7 @@ const deleteSchedule = async (req, res) => {
     return res.status(400).json({
       success: false,
       message:
-        "Cannot delete schedule less than 24 hours before the appointment!",
+        "Bạn không thể xoá lịch làm việc trong vòng 24h trước khi diễn ra!",
     });
   } catch (error) {
     return res.status(500).json({ success: false, message: error.message });
@@ -147,13 +168,46 @@ const getScheduleByDoctorDashboard = async (req, res) => {
   }
 };
 
+// const doctorCreateSchedule = async (req, res) => {
+//   try {
+//     const { id } = req.params;
+//     const { work_shift } = req.body;
+
+//     const doctor = await Doctor.findOne({ user_id: id });
+//     if (!doctor) {
+//       return res.status(400).json({ message: "Doctor not found" });
+//     }
+//     const checkSchedule = await Schedule.findOne({
+//       work_date: req.body.work_date,
+//       doctor_id: doctor._id,
+//     });
+//     if (checkSchedule) {
+//       return res.status(400).json({ message: "Schedule already exists" });
+//     }
+//     const schedule = await Schedule.create({
+//       ...req.body,
+//       doctor_id: doctor._id,
+//     });
+//     if (schedule) {
+//       return res.status(200).json({ success: true, schedule });
+//     }
+//     return res
+//       .status(400)
+//       .json({ success: false, message: "Schedule not found" });
+//   } catch (error) {
+//     return res.status(500).json({ success: false, message: error.message });
+//   }
+// };
 const doctorCreateSchedule = async (req, res) => {
   try {
     const { id } = req.params;
+    const { work_shift, work_date } = req.body;
+
     const doctor = await Doctor.findOne({ user_id: id });
     if (!doctor) {
       return res.status(400).json({ message: "Doctor not found" });
     }
+
     const checkSchedule = await Schedule.findOne({
       work_date: req.body.work_date,
       doctor_id: doctor._id,
@@ -161,16 +215,34 @@ const doctorCreateSchedule = async (req, res) => {
     if (checkSchedule) {
       return res.status(400).json({ message: "Schedule already exists" });
     }
+
+    // Xác định thời gian bắt đầu dựa trên work_shift
+    let startTime;
+
+    if (work_shift === "morning") {
+      startTime = new Date(work_date);
+      startTime.setHours(7, 30, 0, 0); // 7h30
+    } else if (work_shift === "afternoon") {
+      startTime = new Date(work_date);
+      startTime.setHours(13, 30, 0, 0); // 13h30
+    } else {
+      return res.status(400).json({ message: "Invalid work shift" });
+    }
+
+    // Cộng thêm 7 giờ để chuyển đổi sang giờ Việt Nam
+    startTime.setHours(startTime.getHours() + 7);
+
+    // Lưu work_date với giờ đã cập nhật
     const schedule = await Schedule.create({
       ...req.body,
       doctor_id: doctor._id,
+      work_date: startTime.toISOString(), // Lưu work_date với giờ đã cập nhật
     });
+
     if (schedule) {
       return res.status(200).json({ success: true, schedule });
     }
-    return res
-      .status(400)
-      .json({ success: false, message: "Schedule not found" });
+    return res.status(400).json({ success: false, message: "Schedule not found" });
   } catch (error) {
     return res.status(500).json({ success: false, message: error.message });
   }
@@ -291,6 +363,13 @@ const doctorUpdateSchedule = async (req, res) => {
 
       try {
         await transporter.sendMail(mailOptions);
+        await Notification.create({
+          patient_id: appointment.patient_id,
+          doctor_id: appointment.doctor_id,
+          message: `Lịch hẹn của bản thay đổi từ ngày ${formattedOldDate} - Ca khám: ${oldShift} thành ngày ${formattedNewDate} - Ca khám: ${newShift}.`,
+          appointment_id: appointment._id,
+          recipientType: "patient",
+        });
       } catch (emailError) {
         console.error(`Failed to send email to ${userInfo.email}:`, emailError);
       }
